+++ examples/lib/stories/system/system.dart
@@ -2,6 +2,7 @@ import 'package:dashbook/dashbook.dart';
 import 'package:examples/commons/commons.dart';
 import 'package:examples/stories/system/overlays_example.dart';
 import 'package:examples/stories/system/pause_resume_example.dart';
+import 'package:examples/stories/system/resize_example.dart';
 import 'package:examples/stories/system/step_engine_example.dart';
 import 'package:examples/stories/system/without_flame_game_example.dart';
 import 'package:flame/game.dart';
@@ -31,5 +32,11 @@ void addSystemStories(Dashbook dashbook) {
       (_) => GameWidget(game: StepEngineExample()),
       codeLink: baseLink('system/step_engine_game.dart'),
       info: StepEngineExample.description,
+    )
+    ..add(
+      'On Game Resize',
+      (_) => GameWidget(game: ResizeExampleGame()),
+      codeLink: baseLink('system/resize_example.dart'),
+      info: ResizeExampleGame.description,
     );
 }