+++ packages/flame_test/lib/flame_test.dart
@@ -1,5 +1,6 @@
 export 'src/close_to_aabb.dart' show closeToAabb;
 export 'src/close_to_vector.dart';
+export 'src/close_to_vector3.dart';
 export 'src/debug_text_renderer.dart' show DebugTextRenderer;
 export 'src/expect_double.dart';
 export 'src/fails_assert.dart';